requirements.txt
psycopg[binary]>=3.1.18
python-dotenv>=1.0.1
google-cloud-aiplatform>=1.38.0,<2.0.0
vertexai